Collapsed test fixtures into base class
